requirements.txt
Django==1.9.5
Markdown==2.6.6
PyJWT==1.4.0
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
boto==2.40.0
celery==3.1.23
django-celery==3.1.17
django-filter==0.13.0
django-storages==1.4.1
djangorestframework==3.3.3
djangorestframework-jwt==1.8.0
gunicorn==19.4.5
kombu==3.0.35
psycopg2==2.6.1
pytz==2016.4
redis==2.10.5
wsgiref==0.1.2
